requirements/base.txt
amqp
anyjson
billiard
celery
Django==2.2.9
django-admin-sortable2
django-adminplus
django-audit-log
django-bootstrap4
django-ckeditor
django-countries
django-ordered-model
django-webpack-loader
django-simple-captcha
django-tables2
djangorestframework
futures
html5lib
httplib2
kombu
M2Crypto
mercadopago
Pillow
psycopg2-binary
pyaml
pyPdf
PyPDF2
python-dateutil
python-stdnum
pytz
PyYAML
redis
redlock-py
reportlab
requests
simplejson
six
suds-jurko
validate-email
vatnumber
wheel
xhtml2pdf
